+++ src/components/sideBar/sideBar.css
@@ -1,0 +1,320 @@
+/* main */
+* {
+  box-sizing: border-box;
+}
+
+a {
+  display: block;
+  text-decoration: none;
+}
+
+img {
+  display: block;
+  max-width: 100%;
+  height: auto;
+}
+
+*::before,
+*::after {
+  box-sizing: inherit;
+}
+
+ul,
+ol {
+  list-style: none;
+}
+
+/* SIDEBAR */
+
+/* .menu_btn-lines {
+  width: 15px;
+  height: 15px;
+} */
+
+/* .menu_btn-lines,
+.menu_btn-lines::before,
+.menu_btn-lines::after {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  height: 2px;
+  margin-top: -1px;
+  width: 14px;
+  margin-left: -7px;
+  background-color: #000000;
+} */
+
+/* .menu_btn-lines::before,
+.menu_btn-lines::after {
+  content: '';
+  display: block;
+  transition: 1000ms;
+} */
+
+/* .menu_btn-lines::before {
+  transform: translateY(-5px);
+} */
+
+/* .menu_btn-lines::after {
+  transform: translateY(5px);
+} */
+
+/* .menu_btn-active {
+  display: none;
+  width: 100%;
+  height: 100vh;
+  border: 1px #e6e8e8 solid;
+  position: relative;
+  background-color: #f4f6fb;
+} */
+
+/* .menu_btn-active .menu_btn-lines::before {
+  transform: rotate(45deg);
+} */
+
+/* .menu_btn-active .menu_btn-lines {
+  height: 0;
+  top: 10%;
+  left: 90%;
+} */
+
+/* .menu_btn-active .menu_btn-lines::after {
+  transform: rotate(-45deg);
+} */
+
+/* .menu_btn-active,
+.menu_btn-active::before,
+.menu_btn-active::after {
+  position: absolute;
+  top: 0;
+  left: 0;
+} */
+
+/* .wrapper {
+  position: relative;
+  overflow-x: hidden;
+} */
+
+/* .navigation__menu {
+  position: fixed;
+  left: 0;
+  top: 0;
+  width: 90%;
+  height: 100vh;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+} */
+
+.menu-wrapper-none {
+  display: none;
+}
+
+.menu-wrapper {
+  position: fixed;
+  left: 0;
+  top: 0;
+  width: 100%;
+  height: 100vh;
+  background: linear-gradient(
+    to right,
+    #f4f6fb 0%,
+    #f4f6fb 97%,
+    #0000009f 97%,
+    #0000009f 100%
+  );
+  animation: left 600ms ease-out;
+}
+
+@keyframes left {
+  0% {
+    transform: translateX(-100px);
+    opacity: 0;
+  }
+
+  100% {
+    transform: translateX(0);
+    opacity: 1;
+  }
+}
+
+.close {
+  position: absolute;
+  right: 10px;
+  top: 0px;
+  width: 65px;
+  height: 70px;
+  background-color: transparent;
+  border: none;
+  font-size: 24px;
+}
+
+.close:hover,
+.close:focus {
+  color: #ff6b08;
+  font-size: 32px;
+}
+
+.navigation__menu_list {
+  display: flex;
+  justify-content: center;
+  align-items: flex-start;
+  height: 100%;
+  width: 90%;
+  flex-direction: column;
+  top: 70px;
+  left: 0;
+  padding-left: 32px;
+  position: absolute;
+}
+
+.loginOnMobile {
+  display: none;
+}
+
+.navigation__menu_list-item {
+  display: flex;
+  align-items: center;
+  flex-wrap: wrap;
+  height: 65px;
+  width: 100%;
+  font-family: 'Roboto', sans-serif;
+  position: relative;
+  animation: light-SpeedIn 1s ease-out;
+}
+
+@keyframes light-SpeedIn {
+  from {
+    transform: translate3d(-100%, 0, 0) skewX(-30deg);
+    opacity: 0;
+  }
+
+  60% {
+    transform: skewX(20deg);
+    opacity: 1;
+  }
+
+  80% {
+    transform: skewX(-5deg);
+  }
+
+  to {
+    transform: translate3d(0, 0, 0);
+  }
+}
+
+.navigation__menu_list-item-link:hover,
+.navigation__menu_list-item-link:focus {
+  color: #ff6b08;
+  font-size: 20px;
+}
+
+.light-SpeedIn {
+  animation-timing-function: ease-out;
+}
+
+.navigation__menu_list-item:after {
+  content: '';
+  height: 1px;
+  width: 100%;
+  background-color: lightgray;
+  margin-top: 20px;
+  position: absolute;
+}
+
+.navigation__menu_list-item:first-child::before {
+  content: '';
+  height: 1px;
+  width: 100%;
+  background-color: lightgray;
+  margin-top: -21px;
+  position: absolute;
+}
+
+.navigation__menu_list-item-link {
+  text-decoration: none;
+  font-size: 16px;
+  letter-spacing: 0.02em;
+  font-weight: bold;
+  color: #000000;
+}
+
+@media (min-width: 768px) {
+  .loginOnMobile {
+    text-decoration: none;
+    font-size: 16px;
+    letter-spacing: 0.02em;
+    font-weight: bold;
+    color: #000000;
+  }
+
+  .menu-wrapper {
+    background: linear-gradient(
+      to right,
+      #f4f6fb 0%,
+      #f4f6fb 42%,
+      #0000009f 42%,
+      #0000009f 100%
+    );
+  }
+
+  .close {
+    right: 445px;
+  }
+
+  .navigation__menu_list {
+    width: 41%;
+  }
+}
+
+@media (min-width: 1024px) {
+  .menu-wrapper {
+    background: linear-gradient(
+      to right,
+      #f4f6fb 0%,
+      #f4f6fb 33%,
+      #0000009f 33%,
+      #0000009f 100%
+    );
+  }
+
+  .close {
+    right: 685px;
+  }
+
+  .navigation__menu_list {
+    width: 32%;
+  }
+}
+
+@media (min-width: 1280px) {
+  .menu-wrapper {
+    background: linear-gradient(
+      to right,
+      #f4f6fb 0%,
+      #f4f6fb 25%,
+      #0000009f 25%,
+      #0000009f 100%
+    );
+  }
+
+  .close {
+    right: 960px;
+  }
+
+  .navigation__menu_list {
+    width: 24%;
+  }
+}
+
+@media (min-width: 1440px) {
+  .close {
+    right: 1080px;
+  }
+}
+
+@media (min-width: 2560px) {
+  .close {
+    right: 1920px;
+  }
+}
