+++ src/components/templates/menuList.hbs
@@ -1,0 +1,7 @@
+{{#each this}}
+<li class="navigation__menu_list-item">
+    <a href="{{_id}}" class="navigation__menu_list-item-link">
+        {{category}}
+    </a>
+</li>
+{{/each }}
