--- src/components/templates/template.hbs
@@ -1,7 +1,0 @@
-{{#each this}}
-<li class="navigation__menu_list-item">
-    <a href="#idOfCategory" id="{{_id}}" class="navigation__menu_list-item-link">
-        {{category}}
-    </a>
-</li>
-{{/each }}
